Add hook for handling colours.
