+++ src/app/components/series/edit-series/edit-series-modal.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { Colour, buttonHoverColours } from "@/app/components/util";
-import { ChangeEvent, useState } from "react";
 import { PencilIcon } from "../../icons";
 import EditSeriesForm from "./edit-series-form";
 
@@ -12,16 +11,8 @@ interface Props {
 }
 
 const EditSeriesModal = ({ id, name, colour }: Props) => {
-  const [selectedColour, setSelectedColour] = useState<Colour>(colour);
-
   const btnHoverColour = buttonHoverColours[colour];
   const modalId = "edit_series_modal_" + id;
-
-  const handleColourChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    const value = e.target.value;
-    const colour = value.charAt(0).toLowerCase() + value.slice(1);
-    setSelectedColour(colour as Colour);
-  };
 
   return (
     <>
@@ -44,8 +35,6 @@ const EditSeriesModal = ({ id, name, colour }: Props) => {
               id={id}
               initialName={name}
               initialColour={colour}
-              selectedColour={selectedColour}
-              handleColourChange={handleColourChange}
             />
           </div>
         </div>
